+++ backend/src/main/java/com/together/backend/domain/calendar/model/request/CalendarRecordRequest.java
@@ -1,14 +1,16 @@
-package com.together.backend.domain.calendar.dto;
+package com.together.backend.domain.calendar.model.request;
 
 import com.together.backend.domain.calendar.model.entity.CondomUsage;
 import com.together.backend.domain.calendar.model.entity.MoodType;
 import lombok.*;
 
+// 캘린더 기록 등록용 DTO
+
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
-public class CalendarDetailResponse {
+public class CalendarRecordRequest {
     private String date;
+    private Boolean hadSex;
     private Boolean takenPill;
     private CondomUsage usedCondom;
-    private Boolean hadSex;
     private MoodType moodEmoji;
 }
