+++ backend/src/main/java/com/together/backend/domain/calendar/model/response/CalendarDetailResponse.java
@@ -1,15 +1,14 @@
-package com.together.backend.domain.calendar.dto;
+package com.together.backend.domain.calendar.model.response;
 
 import com.together.backend.domain.calendar.model.entity.CondomUsage;
 import com.together.backend.domain.calendar.model.entity.MoodType;
 import lombok.*;
 
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
-public class CalendarRecordResponse {
+public class CalendarDetailResponse {
     private String date;
-    private Boolean hadSex;
     private Boolean takenPill;
-    private CondomUsage usedCondom; // Enum!
+    private CondomUsage usedCondom;
+    private Boolean hadSex;
     private MoodType moodEmoji;
 }
-
