+++ backend/src/main/java/com/together/backend/domain/calendar/model/response/CalendarSummaryResponse.java
@@ -1,4 +1,4 @@
-package com.together.backend.domain.calendar.dto;
+package com.together.backend.domain.calendar.model.response;
 
 import com.together.backend.domain.calendar.model.entity.MoodType;
 import lombok.*;
