+++ backend/src/main/java/com/together/backend/domain/notification/service/NotificationService.java
@@ -3,11 +3,11 @@ package com.together.backend.domain.notification.service;
 import com.together.backend.domain.calendar.model.entity.IntakeRecord;
 import com.together.backend.domain.calendar.repository.IntakeRecordRepository;
 import com.together.backend.domain.notification.model.Notification;
-import com.together.backend.domain.notification.model.NotificationSettings;
+import com.together.backend.domain.notificationSettings.model.NotificationSettings;
 import com.together.backend.domain.notification.model.NotificationType;
 import com.together.backend.domain.notification.repository.NotificationRepository;
-import com.together.backend.domain.notification.repository.NotificationSettingsRepository;
-import com.together.backend.domain.pill.model.UserPill;
+import com.together.backend.domain.notificationSettings.repository.NotificationSettingsRepository;
+import com.together.backend.domain.pill.model.entity.UserPill;
 import com.together.backend.domain.pill.repository.UserPillRepository;
 import com.together.backend.domain.user.model.entity.User;
 import com.together.backend.domain.user.repository.UserRepository;
