+++ backend/src/main/java/com/together/backend/domain/notificationSettings/repository/NotificationSettingsRepository.java
@@ -1,11 +1,10 @@
-package com.together.backend.domain.notification.repository;
+package com.together.backend.domain.notificationSettings.repository;
 
-import com.together.backend.domain.notification.model.NotificationSettings;
+import com.together.backend.domain.notificationSettings.model.NotificationSettings;
 import com.together.backend.domain.notification.model.NotificationType;
 import com.together.backend.domain.user.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
