+++ backend/src/main/java/com/together/backend/domain/pill/model/entity/IntakeOption.java
@@ -1,4 +1,4 @@
-package com.together.backend.domain.pill.model;
+package com.together.backend.domain.pill.model.entity;
 
 import lombok.Getter;
