+++ backend/src/main/java/com/together/backend/domain/pill/repository/IntakeInfoRepository.java
@@ -1,6 +1,6 @@
 package com.together.backend.domain.pill.repository;
 
-import com.together.backend.domain.pill.model.IntakeInfo;
+import com.together.backend.domain.pill.model.entity.IntakeInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
