+++ backend/src/main/java/com/together/backend/domain/pill/service/UserPillService.java
@@ -4,12 +4,12 @@ import com.together.backend.domain.calendar.model.entity.IntakeRecord;
 import com.together.backend.domain.calendar.repository.BasicRecordRepository;
 import com.together.backend.domain.calendar.repository.IntakeRecordRepository;
 import com.together.backend.domain.calendar.service.IntakeRecordInitService;
-import com.together.backend.domain.notification.model.NotificationSettings;
+import com.together.backend.domain.notificationSettings.model.NotificationSettings;
 import com.together.backend.domain.notification.model.NotificationType;
-import com.together.backend.domain.notification.repository.NotificationSettingsRepository;
-import com.together.backend.domain.pill.model.IntakeInfo;
-import com.together.backend.domain.pill.model.IntakeOption;
-import com.together.backend.domain.pill.model.UserPill;
+import com.together.backend.domain.notificationSettings.repository.NotificationSettingsRepository;
+import com.together.backend.domain.pill.model.entity.IntakeInfo;
+import com.together.backend.domain.pill.model.entity.IntakeOption;
+import com.together.backend.domain.pill.model.entity.UserPill;
 import com.together.backend.domain.pill.model.request.UserPillRequest;
 import com.together.backend.domain.pill.model.response.TodayPillResponse;
 import com.together.backend.domain.pill.model.response.UserPillRemainResponse;
