+++ backend/src/main/java/com/together/backend/domain/sharing/repository/SharingRepository.java
@@ -1,6 +1,6 @@
 package com.together.backend.domain.sharing.repository;
 
-import com.together.backend.domain.sharing.model.Sharing;
+import com.together.backend.domain.sharing.model.entity.Sharing;
 import com.together.backend.domain.user.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
